[Fix] Qna CustomUserDetails 서비스 전달 방식 수정 및 목록 조회시 super카테고리 예외처리

+++ backend/src/main/java/org/example/backend/Qna/Controller/QuestionController.java
@@ -37,7 +37,7 @@ public class QuestionController {
 
     //qna 상세 조회
     @GetMapping("/detail")
-    public BaseResponse<GetQuestionDetailRes> getQnaDetail(Long qnaBoardId) {
+    public BaseResponse<GetQuestionDetailRes> getQnaDetail(Integer qnaBoardId) {
         GetQuestionDetailRes questionDetailRes = qnaService.getQuestionDetail(qnaBoardId);
         return new BaseResponse<>(questionDetailRes);
